Added debugging module
