+++ pennylane/debugging/__init__.py
@@ -1,0 +1,28 @@
+# Copyright 2018-2022 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+This module contains functionality for debugging quantum programs on simulator devices.
+"""
+
+from .snapshots import snapshots
+
+from .debugger import (
+    breakpoint, 
+    expval,
+    PLDB,
+    pldb_device_manager,
+    probs,
+    state,
+    tape,
+)
